# BACK/swagger.yaml
openapi: 3.0.0
info:
  title: API Doceria Gourmet IANES
  description: >
    API RESTful para gerenciamento de estoque da Doceria Gourmet IANES.
    Controla produtos, lotes, movimentações e notificações de validade e estoque baixo.
  version: 1.0.0
  contact:
    name: Sidney Campos Aragão
servers:
  - url: http://localhost:8080/v1/IANES
    description: Ambiente de desenvolvimento
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErroValidacao:
      type: object
      properties:
        status:       { type: string,  example: "Bad Request" }
        status_code:  { type: integer, example: 400 }
        message:      { type: string,  example: "Campos obrigatórios não preenchidos corretamente." }
    ErroNaoEncontrado:
      type: object
      properties:
        status:       { type: string,  example: "Not Found" }
        status_code:  { type: integer, example: 404 }
        message:      { type: string,  example: "Nenhum item encontrado." }
    ErroNaoAutorizado:
      type: object
      properties:
        status:       { type: string,  example: "Unauthorized" }
        status_code:  { type: integer, example: 401 }
        message:      { type: string,  example: "Token inválido ou expirado." }
...
paths:
...
  /usuario:
    get:
      tags: [Usuário]
      summary: Listar todos os usuários
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Lista de usuários retornada com sucesso
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Requisição realizada com sucesso!"
                response:
                  usuarios:
                    - id_usuario: 1
                      nome: "Sidney Aragão"
                      email: "sidney@ianes.com"
                      is_active: true
                    - id_usuario: 2
                      nome: "Maria Doces"
                      email: "maria@ianes.com"
                      is_active: true
        '401': { description: Token ausente ou inválido, content: { application/json: { schema: { $ref: '#/components/schemas/ErroNaoAutorizado' } } } }
        '404': { description: Nenhum usuário encontrado, content: { application/json: { schema: { $ref: '#/components/schemas/ErroNaoEncontrado' } } } }

    post:
      tags: [Usuário]
      summary: Cadastrar novo usuário
      requestBody:
        required: true
        content:
          application/json:
            example:
              nome: "Sidney Aragão"
              email: "sidney@ianes.com"
              senha: "Senai@2026"
      responses:
        '201':
          description: Usuário cadastrado com sucesso
          content:
            application/json:
              example:
                status: "Created"
                status_code: 201
                message: "Item criado com sucesso!"
                response:
                  usuario:
                    id_usuario: 1
                    nome: "Sidney Aragão"
                    email: "sidney@ianes.com"
        '400': { description: Dados inválidos ou email já cadastrado, content: { application/json: { schema: { $ref: '#/components/schemas/ErroValidacao' } } } }
  /usuario/{id}:
    get:
      tags: [Usuário]
      summary: Buscar usuário por ID
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, example: 1 }
      responses:
        '200':
          description: Usuário encontrado
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Requisição realizada com sucesso!"
                response:
                  usuario:
                    id_usuario: 1
                    nome: "Sidney Aragão"
                    email: "sidney@ianes.com"
                    is_active: true
        '404': { description: Usuário não encontrado, content: { application/json: { schema: { $ref: '#/components/schemas/ErroNaoEncontrado' } } } }

    put:
      tags: [Usuário]
      summary: Atualizar dados do usuário (parcial)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, example: 1 }
      requestBody:
        required: true
        content:
          application/json:
            example:
              nome: "Sidney Campos"
      responses:
        '200':
          description: Usuário atualizado com sucesso
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Item atualizado com sucesso!"
                response:
                  usuario:
                    id_usuario: 1
                    nome: "Sidney Campos"
        '400': { description: Dados inválidos, content: { application/json: { schema: { $ref: '#/components/schemas/ErroValidacao' } } } }
        '404': { description: Usuário não encontrado, content: { application/json: { schema: { $ref: '#/components/schemas/ErroNaoEncontrado' } } } }

    delete:
      tags: [Usuário]
      summary: Desativar usuário (soft delete)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, example: 1 }
      responses:
        '200':
          description: Usuário desativado com sucesso
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Item deletado com sucesso!"
        '404': { description: Usuário não encontrado, content: { application/json: { schema: { $ref: '#/components/schemas/ErroNaoEncontrado' } } } }
  /usuario/login:
    post:
      tags: [Usuário]
      summary: Realizar login e obter Bearer Token
      requestBody:
        required: true
        content:
          application/json:
            example:
              email: "sidney@ianes.com"
              senha: "Senai@2026"
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Requisição realizada com sucesso!"
                response:
                  usuario:
                    id_usuario: 1
                    nome: "Sidney Aragão"
                    email: "sidney@ianes.com"
                  token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              example:
                status: "Unauthorized"
                status_code: 401
                message: "Email ou senha incorretos."
  /usuario/recuperar-senha:
    post:
      tags: [Usuário]
      summary: Solicitar código de recuperação de senha por e-mail
      requestBody:
        required: true
        content:
          application/json:
            example:
              email: "sidney@ianes.com"
      responses:
        '200':
          description: Solicitação processada (resposta genérica intencional)
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Se o e-mail estiver cadastrado, você receberá o código em breve."
  /usuario/redefinir-senha:
    post:
      tags: [Usuário]
      summary: Redefinir senha com código de 6 dígitos recebido por e-mail
      requestBody:
        required: true
        content:
          application/json:
            example:
              email: "sidney@ianes.com"
              codigo: "483921"
              nova_senha: "NovaSenha@2026"
      responses:
        '200':
          description: Senha redefinida com sucesso
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Senha redefinida com sucesso!"
        '400':
          description: Código inválido ou expirado
          content:
            application/json:
              example:
                status: "Bad Request"
                status_code: 400
                message: "Campos obrigatórios não preenchidos corretamente. [codigo inválido ou expirado]"

  # ══════════════════════════════════════════════
  # CATEGORIA
  # ══════════════════════════════════════════════
...